requirements.txt
google-adk
pandas
uvicorn
requests
